Refactor #92: split DocumentPreview into TOC and node-renderer modules
Extract the table-of-contents (buildTocTree + PreviewToc/TocList/TocListItem)
into PreviewToc.tsx and the per-format node renderers (PreviewNode, HeadingNode,
ContentNode, ListNode, ListItemNode, FootnoteSection + markup helpers) into
PreviewNodeRenderers.tsx. DocumentPreview.tsx is now a ~50-line orchestrator
(layout, outline memo, footnote split).

Pure extraction — moved code is byte-for-byte identical apart from added export
keywords. Adds unit tests for buildTocTree and each renderer in isolation; the
existing DocumentPreview suite stays green as the regression guard.
